requirements.txt
# AI Models
anthropic==0.39.0
openai==1.54.3
google-generativeai==0.8.3

# LangChain
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-anthropic==0.3.0

# Vector DB
qdrant-client==1.12.1

# Database
supabase==2.9.0
psycopg2-binary==2.9.10
redis==5.2.0
sqlalchemy==1.4,<2.1

# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic>=2.12.0
pydantic-settings==2.6.0

# MCP (Model Context Protocol)
mcp[cli]>=1.20.0

# Utilities
httpx==0.27.2
aiohttp==3.11.2
tenacity==9.0.0
python-dotenv==1.0.1
rich==13.9.4
click==8.1.7
pyyaml==6.0.2

# Auth
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# Monitoring & Error Tracking
sentry-sdk[fastapi]==2.0.0
psutil==5.9.8

# Web Scraping (for rankings collector)
beautifulsoup4==4.12.3
requests==2.31.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
